transactions getting from the pool to the sequencer stage
